Fix classcastexception for some type inference tests
This problem surfaced that we are depending too much on JDT not
tidying up. Here the tidyup was done for the TypeSystem object
then we attempted to use it again. I don't have the time to
unravel the subtleties of why we're doing things like this so
did the quick thing and prevented the tidy up. We should keep
an eye on memory and ensure no leaks are occurring when
repeatedly driving type inference.