Fix for #706: do method resolve after verifyMethods for @Override checks
Java steps happen serially for each CompUnitDecl, however Groovy resolve
is run all at once for all CompUnitDecl in the CompUnit.  So the steps
verifyMethods and resolve needed to be called in JDTResolver to ensure
proper sequencing and so a JDT scope is available for Groovy's resolve.

Resolving fields from the same spot resulted in error for enum constants
with statement bodies.