This is part of an experiment to run AST transforms on source references
It was partially working at one point in time, but has fallen behind
some of the recent changes that were made to support anonymous inner
classes.

The idea is to run the AST transforms on source references as soon as
resolve information is available.  In my determination, that point is in
CompilationUnitScope.buildTypeBindings.  I tried to find an earlier
time, but necessary type information was not yet available.  Once
transforms are run, the CompilationUnitDeclaration needs to be
re-processed into the JDT model to make the stuff the AST transforms
have modified available to the Java side of things.

The overall goal is to provide a better joint compilation experience.
If a Groovy class is annotated with @Canonical, a Java class should
expect to have access to the extra constructors.